:arrow_up: Resolve `semver` version to 7.5.3
The current version of `@vitest/coverage-istanbul` relies on an outdated and deprecated version of `istanbul`, which in turn uses `istanbul-lib-instrument` that includes the `semver` package.

Recently, the `semver` package has been found to have a moderate vulnerability, as detailed in the ["semver vulnerable to Regular Expression Denial of Service"](https://github.com/advisories/GHSA-c2qf-rxjj-qqgw) advisory. This vulnerability fails the dependencies audit and poses a security risk.

To address this issue and ensure compliance with security standards, this commit forces the usage of the latest semver version (7.5.3) by updating the dependency in `@vitest/coverage-istanbul`.